feat: separation of model and business logic
